move OperationKind to separate source and test files
